+++ CHANGELOG.md
@@ -5,6 +5,29 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.4.0 - 2021-10-03
+
+# Added
+
+  * Room registration (new ``Room`` and ``Rental`` models).
+  * Room check integrated in registration process.
+  * Rezidence rooms created automatically if table empty.
+  * Connection check page, integrated in registration process (new Bower
+    dependency: webping-js).
+  * Global progress bar in registration process.
+  * Past / Future Date form validators.
+  * New utility function ``get_bootstrap_icon`` to facilite icons integration.
+
+# Changed
+
+  * New favicon
+  * New 404 / other errors pages.
+
+# Fixed
+
+  * Discord report of app errors.
+
+
 ## 0.3.1 - 2021-09-30
 
 # Changed